[booster-silica] Add a login session root parent booster . Contributes to JB#40880
If this booster is launched as a child of a login session and
is used to boost the silica boosters, all boosted graphical
applications will be made part of the login session. Which is
good because it matters to polkit that authenticating processes
and authentication agent processes belong to the same session.